patch: migrated whois -> rdap (#24)
