+++ internal/domain/domain.go
@@ -1,62 +1,24 @@
 package domain
 
 import (
-	"context"
 	"errors"
 	"fmt"
-	"math/rand/v2"
 	"regexp"
 	"strings"
-	"sync"
-	"time"
 
-	"github.com/likexian/whois"
-	whoisparser "github.com/likexian/whois-parser"
+	"github.com/openrdap/rdap"
 	"golang.org/x/net/publicsuffix"
 )
-
-type ConfigOptions struct {
-	TLDs            []string
-	Prefixes        []string
-	Suffixes        []string
-	MaxDomainLength int
-	Verbose         bool
-	OnlyAvailable   bool
-	ShowStats       bool
-}
-
-type DomainResult struct {
-	Domain    string `json:"domain"`
-	Available bool   `json:"available"`
-	Error     error  `json:"error,omitempty"`
-}
-
-type Stats struct {
-	total        int
-	available    int
-	notAvailable int
-	timedOut     int
-	errored      int
-}
-
-const (
-	maxRetries       = 3
-	initialBackoff   = 500 * time.Millisecond
-	backoffFactor    = 5.0
-	jitterFraction   = 0.7 // +/-70% randomness
-	contextTimeout   = 15 * time.Second
-	concurrencyLimit = 20
-)
-
-var Config = ConfigOptions{}
-
-var stats = Stats{}
 
 func Exec(domainsOrKeywords []string) {
 	keywords := validateKeywords(domainsOrKeywords)
 	domains := generateDomainPermutations(keywords)
 	stats.total = len(domains)
-	resultChan := checkDomainsStreaming(domains, concurrencyLimit, contextTimeout)
+	resolverService := ResolverService{
+		config:     &Config,
+		rdapClient: &rdap.Client{},
+	}
+	resultChan := resolverService.checkDomainsStreaming(domains, concurrencyLimit, contextTimeout)
 
 	for result := range resultChan {
 		if result.Error != nil {
@@ -80,108 +42,6 @@ func Exec(domainsOrKeywords []string) {
 	if Config.ShowStats {
 		fmt.Println(RenderStatsSummary())
 	}
-}
-
-func checkAvailability(ctx context.Context, domain string) (bool, error) {
-	if !isValidDomainOrKeyword(domain) {
-		return false, errors.New("Invalid domain")
-	}
-
-	var lastErr error
-	backoff := initialBackoff
-
-	for attempt := 0; attempt <= maxRetries; attempt++ {
-		select {
-		case <-ctx.Done():
-			return false, ctx.Err()
-		default:
-			time.Sleep(time.Duration(float64(backoff) * (1 + (rand.Float64()*2-1)*jitterFraction)))
-			raw, err := whois.Whois(domain)
-			if err != nil || isTransientWhoisError(err, raw) {
-				lastErr = err
-				if attempt < maxRetries {
-					jitter := time.Duration(float64(backoff) * (1 + (rand.Float64()*2-1)*jitterFraction))
-					time.Sleep(jitter)
-					backoff = time.Duration(float64(backoff) * backoffFactor)
-					continue
-				}
-				stats.timedOut += 1
-				return false, fmt.Errorf("whois error after %d retries: %w", attempt, err)
-			}
-
-			parsed, err := whoisparser.Parse(raw)
-			if err != nil && strings.Contains(err.Error(), "domain is not found") {
-				return true, nil
-			}
-			if parsed.Registrar != nil {
-				return false, nil
-			}
-			return false, err
-		}
-	}
-
-	return false, fmt.Errorf("unreachable: exhausted retries, last error: %v", lastErr)
-}
-
-func isTransientWhoisError(err error, raw string) bool {
-	if err == nil && raw == "" {
-		return true // empty response
-	}
-	if err != nil {
-		msg := err.Error()
-		return strings.Contains(msg, "connection reset") ||
-			strings.Contains(msg, "timeout") ||
-			strings.Contains(msg, "EOF") ||
-			strings.Contains(msg, "refused") ||
-			strings.Contains(msg, "too many requests")
-	}
-	return false
-}
-
-func checkDomainsStreaming(domains []string, concurrency int, timeout time.Duration) <-chan DomainResult {
-	resultChan := make(chan DomainResult)
-	inputChan := make(chan string)
-
-	go func() {
-		defer close(inputChan)
-		for _, domain := range domains {
-			inputChan <- domain
-		}
-	}()
-
-	go func() {
-		var wg sync.WaitGroup
-		sem := make(chan struct{}, concurrency)
-
-		for domain := range inputChan {
-			domain := domain
-			sem <- struct{}{}
-			wg.Add(1)
-			time.Sleep(50 * time.Millisecond) // Throttle requests to avoid rate limiting
-
-			go func() {
-				defer func() {
-					<-sem
-					wg.Done()
-				}()
-
-				ctx, cancel := context.WithTimeout(context.Background(), timeout)
-				defer cancel()
-
-				available, err := checkAvailability(ctx, domain)
-				resultChan <- DomainResult{
-					Domain:    domain,
-					Available: available,
-					Error:     err,
-				}
-			}()
-		}
-
-		wg.Wait()
-		close(resultChan)
-	}()
-
-	return resultChan
 }
 
 func generateDomainPermutations(keywords []string) []string {
@@ -288,4 +148,16 @@ func removeDuplicates(strs []string) []string {
 		}
 	}
 	return list
+}
+
+func PrintDomain(domain string, available bool, err error) {
+	if err != nil {
+		fmt.Println(Errored(domain, err))
+		return
+	}
+	if available {
+		fmt.Println(Available(domain))
+	} else {
+		fmt.Println(NotAvailable(domain))
+	}
 }
